+++ changelog.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### 5.2.0
+- Handle deep-cloning of the search request
+- Avoid returning the search request from getFilter and getFilterValue methods
+
 ### 5.1.0
 - Removing filters by name